% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_aspre.R
\name{add_aspre_interactions}
\alias{add_aspre_interactions}
\title{Add various interaction terms to X. Interaction terms correspond to those in ASPRE,}
\usage{
add_aspre_interactions(X)
}
\arguments{
\item{X}{data frame}
}
\value{
New data frame containing interaction terms.
}
\description{
Adds interaction terms corresponding to ASPRE model
}
\examples{

# Lorem Ipsum
}
\keyword{aspre}
