% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_init.R
\docType{data}
\name{aspre_emulation}
\alias{aspre_emulation}
\title{Emulation-based OHS estimation for ASPRE}
\format{
An object of class \code{list} of length 4.
}
\usage{
aspre_emulation
}
\description{
This object contains data relating to emulation-based OHS estimation for the ASPRE model. For generation, see hidden code in vignette, or in pipeline at https://github.com/jamesliley/OptHoldoutSize_pipelines
}
\keyword{aspre}
\keyword{data}
