% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_init.R
\docType{data}
\name{ci_cover_cost_e_yn}
\alias{ci_cover_cost_e_yn}
\title{Data for example on empirical confidence interval for min cost.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 11 rows and 5000 columns.
}
\usage{
ci_cover_cost_e_yn
}
\description{
Data for example for empirical confidence interval for min cost. For generation, see example.
}
\keyword{aspre}
\keyword{data}
