% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_init.R
\docType{data}
\name{data_nextpoint_par}
\alias{data_nextpoint_par}
\title{Data for 'next point' demonstration vignette on algorithm comparison using parametric algorithm}
\format{
An object of class \code{list} of length 6.
}
\usage{
data_nextpoint_par
}
\description{
Data containing 'next point selected' information for parametric algorithm in vignette comparing emulation and parametric algorithms. For generation, see hidden code in vignette, or in pipeline at https://github.com/jamesliley/OptHoldoutSize_pipelines
}
\keyword{aspre}
\keyword{data}
