% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_simulation.R
\name{gen_dr_coefs}
\alias{gen_dr_coefs}
\title{Generate coefficients corresponding to an imperfect risk score, interpretable
as 'baseline' behaviour in the absence of a risk score}
\usage{
gen_dr_coefs(coefs, noise = TRUE, num_vars = 2, max_dr_powers = 1)
}
\arguments{
\item{coefs}{Original coefficients}

\item{noise}{Set to TRUE to add Gaussian noise to coefficients}

\item{num_vars}{Number of variables at hand for baseline calculation}

\item{max_dr_powers}{??}
}
\value{

}
\description{
Coefficients for imperfect risk score
}
\examples{

# Lorem Ipsum
}
\keyword{simulation}
