% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_simulation.R
\name{gen_preds}
\alias{gen_preds}
\title{Generate matrix of random observations}
\usage{
gen_preds(nobs, npreds, ninters = 0)
}
\arguments{
\item{nobs}{Number of observations (samples)}

\item{npreds}{Number of predictors}

\item{ninters}{Number of interaction terms, default 0}
}
\value{
Data frame of observations
}
\description{
Generate matrix of random observations. Observations are unit Gaussian-distributed.
}
\examples{

# Lorem Ipsum
}
\keyword{simulation}
