% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_init.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(x)
}
\arguments{
\item{x}{argument}
}
\value{
value of logit(x)
}
\description{
Logit function: 1/(1+exp(-x))
}
\examples{

# Plot
x=seq(-5,5,length=1000)
plot(x,logit(x),type="l")
}
