% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_simulation.R
\name{model_predict}
\alias{model_predict}
\title{Make predictions}
\usage{
model_predict(
  data_test,
  trained_model,
  return_type,
  threshold = NULL,
  model_family = NULL,
  ...
)
}
\arguments{
\item{data_test}{Data for which predictions are to be computed}

\item{trained_model}{Model for which predictions are to be made}

\item{return_type}{??}

\item{threshold}{??}

\item{model_family}{??}

\item{...}{Passed to function \code{predict.glm()} or \code{predict.ranger()}}
}
\value{
Vector of predictions
}
\description{
Make predictions according to a given model
}
\examples{

# Lorem Ipsum
}
\keyword{simulation}
