% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_init.R
\docType{data}
\name{ohs_resample}
\alias{ohs_resample}
\title{Data for vignette on algorithm comparison}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 4 columns.
}
\usage{
ohs_resample
}
\description{
This object contains data relating to the first plot in the vignette comparing emulation and parametric algorithms. For generation, see hidden code in vignette, or in pipeline at https://github.com/jamesliley/OptHoldoutSize_pipelines
}
\keyword{aspre}
\keyword{data}
