% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_estimation.R
\name{plot.optholdoutsize_emul}
\alias{plot.optholdoutsize_emul}
\title{Plot estimated cost function using emulation (semiparametric)}
\usage{
\method{plot}{optholdoutsize_emul}(X, k2 = powerlaw, ...)
}
\arguments{
\item{X}{Object of type \code{optholdoutsize_emul}}

\item{k2}{Function governing expected cost to an individual sample given a predictive score fitted to n samples. Must take two arguments: n (number of samples) and theta (parameters). Defaults to a power-law form \verb{powerlaw(n,c(a,b,c))=a n^(-b) + c}.}

\item{...}{Other arguments passed to \code{plot()}}
}
\description{
Plot estimated cost function, when semiparametric (emulation) method is used for estimation.

Draws posterior mean of cost function as a line and indicates minimum. Also draws mean +/- 3 SE.

Assumes a power-law form of k2 unless parameter k2 is set otherwise.
}
\examples{

# Simple example

# Parameters
N=100000;
k1=0.3
A=8000; B=1.5; C=0.15; theta=c(A,B,C)

# True mean function
k2_true=function(n) powerlaw(n,theta)

# Values of n for which cost has been estimated
np=50 # this many points
nset=round(runif(np,1,N))
var_w=runif(np,0.001,0.002)
d=rnorm(np,mean=k2_true(nset),sd=sqrt(var_w))

# Compute OHS
res1=optimal_holdout_size_emulation(nset,d,var_w,N,k1)

# Plot
plot(res1)
}
\keyword{estimation}
