% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_aspre.R
\name{sens10}
\alias{sens10}
\title{Computes sensitivity of a risk score at a threshold at which
10\% of samples (or some proportion pi_int) are above it.}
\usage{
sens10(Y, Ypred, pi_int = pi_intervention)
}
\arguments{
\item{Y}{True labels}

\item{Ypred}{Predictions}

\item{pi}{Compute sensitivity when a proportion pi_int of samples exceed threshold, default 0.1}
}
\value{
Sensitivity at this threshold
}
\description{
Sensitivity at theshold quantile 10\%
}
\examples{

# Lorem Ipsum
}
\keyword{aspre}
