% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_aspre.R
\name{sim_random_aspre}
\alias{sim_random_aspre}
\title{Generate random population of parameters where properties match that of params
Assume independence of parameter variation for simplicity.}
\usage{
sim_random_aspre(n, params = params_aspre)
}
\arguments{
\item{n}{size of population}

\item{pars}{list of parameters}
}
\value{

}
\description{
Simulate random dataset similar to ASPRE training data
}
\examples{

# Load ASPRE related data
data(params_aspre)

# etc
}
\keyword{aspre}
