% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_aspre.R
\name{add_aspre_interactions}
\alias{add_aspre_interactions}
\title{Add interaction terms corresponding to ASPRE model}
\usage{
add_aspre_interactions(X)
}
\arguments{
\item{X}{data frame}
}
\value{
New data frame containing interaction terms.
}
\description{
Add various interaction terms to X. Interaction terms correspond to those in ASPRE.
}
\examples{

# Load ASPRE related data
data(params_aspre)

X=sim_random_aspre(1000,params_aspre)
Xnew=add_aspre_interactions(X)

print(colnames(X))
print(colnames(Xnew))

}
\keyword{aspre}
