% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptHoldoutSize_aspre.R
\name{powersolve}
\alias{powersolve}
\title{Fit power law curve
Find least-squares solution: MLE of (a,b,c) under model
y= ax^-b + c + e, e~N(0,s^2)}
\usage{
powersolve(x, y, init = c(1, -0.8, 0.0165), ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{init}{initial values of (a,b,c) to start}

\item{...}{further parameters passed to optim}
}
\value{
MLE values of (a,b,c)
}
\description{
Fit power law curve
}
\examples{

# Lorem Ipsum
}
\keyword{aspre}
